Split the the code and use ncc (#11)
The separate executable file is unnecessary as we can just easily
combine both the executable and a the library main in the single
file. This will enable us to minify and pack `Micri` into a
single file once we find a feasible solution for that.

- Splitting the code improves readability and maintainability.
- Create single entry point to Micri
- Also remove undocumented deprecated args.
- Remove deprecated options from the parser
-  Build with ncc

This reduces the package size significantly.