Update "iconv-lite" to v0.6.3
There is an issue with "iconv-lite" v0.4.x such that it invokes the
deprecated `new Buffer()` function, which causes Node.js to print an
ugly deprecation warning at runtime.

The `raw-body` dep is the actual consumer of the iconv module, however
even its latest version still uses the problematic iconv version. So
switching to a fork that I published with only that version update
change.